+++ src/tasks.ts
@@ -9,7 +9,7 @@ import FeedProcessor from './services/FeedProcessor.js';
 import loadFeeds from './services/loadFeeds.js';
 import Settings from './types/Settings.js';
 import Block from './models/Block.js';
-import PolygonIOPriceInitializer from './services/PolygonIOPriceInitializer.js';
+import PolygonIOPriceInitializer from './services/fetchers/common/PolygonIOPriceInitializer.js';
 import TimeService from './services/TimeService.js';
 import Blockchain from './lib/Blockchain.js';